requirements.txt
Django==2.0.5
djangorestframework==3.8.2
coverage==4.5.1